% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gevrProfShape.R
\name{gevrProfShape}
\alias{gevrProfShape}
\title{GEVr Shape Parameter Profile Likelihood Estimation}
\usage{
gevrProfShape(z, conf = 0.95, opt = c("Nelder-Mead", "SANN", "BFGS", "CG",
  "L-BFGS-B", "Brent"))
}
\arguments{
\item{z}{A class object returned from gevrFit.}

\item{conf}{Confidence level to use. Defaults to 95 percent.}

\item{opt}{Optimization method to maximize the profile likelihood, passed to optim. The default method is Nelder-Mead.}
}
\value{
Estimate Estimated shape parameter.

CI Profile likelihood based confidence interval for the shape parameter.

ConfLevel The confidence level used.
}
\description{
Computes the profile likelihood based confidence interval for the shape parameter of the GEVr model.
}
\examples{
## Compare the length of the shape confidence intervals using GEV1 vs. GEV10
set.seed(7)
x <- rgevr(200, 10, loc = 0.5, scale = 1, shape = 0.3)
z1 <- gevrFit(x[, 1])
z2 <- gevrFit(x)
gevrProfShape(z1)
gevrProfShape(z2)
}

