% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gevrSeqTests.R
\name{gevrSeqTests}
\alias{gevrSeqTests}
\title{Sequential Tests for the GEVr Model}
\usage{
gevrSeqTests(data, method = "ed", ...)
}
\arguments{
\item{data}{Data should be contain n rows, each a GEVr observation.}

\item{method}{string indicating the method. Only \code{ed} is supported.}

\item{...}{currently ignored (unused arguments) for compatibility with former versions of the package}
}
\value{
Function returns a data frame containing the test statistics, estimates, and p-value results of the sequential tests.

\item{r}{Value of r to be tested.}
\item{p.values}{Raw p-values from the individual tests at each value of r.}
\item{ForwardStop}{Transformed p-values according to the ForwardStop stopping rule.}
\item{StrongStop}{Transformed p-values according to the StrongStop stopping rule.}
\item{statistic}{Returned test statistics of each individual test.}
\item{est.loc}{Estimated location parameter for the given r.}
\item{est.scale}{Estimated scale parameter for the given r.}
\item{est.shape}{Estimated shape parameter for the given r.}
}
\description{
Sequentially performs the entropy difference (ED) test for the GEVr model.
}
\details{
GEVr data (in matrix x) should be of the form \eqn{x[i,1] > x[i, 2] > \cdots > x[i, r]} for each observation \eqn{i = 1, \ldots, n}.
See function `pSeqStop' for details on transformed p-values.
}
\examples{
x <- rgevr(200, 5, loc = 0.5, scale = 1, shape = 0.25)
gevrSeqTests(x, method = "ed")
}
