% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpdProfShape.R
\name{gpdProfShape}
\alias{gpdProfShape}
\title{GPD Shape Parameter Profile Likelihood Estimation for Stationary Models}
\usage{
gpdProfShape(z, conf = 0.95, plot = TRUE)
}
\arguments{
\item{z}{A class object returned from gpdFit.}

\item{conf}{Confidence level to use. Defaults to 95 percent.}

\item{plot}{Plot the profile likelihood and estimate (vertical line)?}
}
\value{
\item{Estimate}{Estimated shape parameter.}
\item{CI}{Profile likelihood based confidence interval for the shape parameter.}
\item{ConfLevel}{The confidence level used.}
}
\description{
Computes the profile likelihood based confidence interval for the shape parameter of the stationary Generalized Pareto model.
}
\examples{
x <- rgpd(200, loc = 0, scale = 1, shape = 0.25)
z <- gpdFit(x, threshold = 0)
gpdProfShape(z)
}
